+++ py/desitarget/cmx/cmx_cuts.py
@@ -2,32 +2,31 @@
 desitarget.cmx.cmx_cuts
 ========================
 
-Target Selection for DESI commissioning (cmx)
-
-https://desi.lbl.gov/trac/wiki/TargetSelectionWG/CommissioningTargets
+Target Selection for DESI commissioning (cmx) derived from `the wiki`_.
 
 A collection of helpful (static) methods to check whether an object's
-flux passes a given selection criterion (*e.g.* STD_DITHER).
+flux passes a given selection criterion (*e.g.* STD_TEST).
+
+.. _`the Gaia data model`: https://gea.esac.esa.int/archive/documentation/GDR2/Gaia_archive/chap_datamodel/sec_dm_main_tables/ssec_dm_gaia_source.html
+.. _`the wiki`: https://desi.lbl.gov/trac/wiki/TargetSelectionWG/CommissioningTargets
 """
-import warnings
+
 from time import time
-import os.path
-
-import numbers
-import sys
-
 import numpy as np
-from pkg_resources import resource_filename
+import os
+
+import warnings
 
 import astropy.units as u
 from astropy.coordinates import SkyCoord
 from astropy.table import Table, Row
 
-from desitarget import io
-from desitarget.internal import sharedmem
-from desitarget.cmx.cmx_targetmask import cmx_mask
 
-from desitarget.gaiamatch import match_gaia_to_primary, pop_gaia_coords
+from desitarget import io
+from desitarget.cuts import _psflike, _is_row
+from desitarget.internal import sharedmem
+from desitarget.targets import finalize
+from desitarget.cmx.cmx_targetmask import cmx_mask
 
 #ADM set up the DESI default logger
 from desiutil.log import get_logger
@@ -37,185 +36,300 @@ log = get_logger()
 start = time()
 
 
-def isSTD_dither(gflux=None):
-    """Placeholder for Gaia dithering targets
-       
-    Args:
-        gflux
-            The flux in nano-maggies of g
-    
-    Returns:
-        mask : array_like. True if and only if the object is Gaia
-            dithering target.
-    """
+def passesSTD_logic(gfracflux=None, rfracflux=None, zfracflux=None,
+                    objtype=None, isgaia=None, pmra=None, pmdec=None,
+                    aen=None, dupsource=None, paramssolved=None,
+                    primary=None):
+    """The default logic/mask cuts for commissioning stars.
 
-    isdither &= (gflux > 0)
+    Parameters
+    ----------
+    gfracflux, rfracflux, zfracflux : :class:`array_like` or :class:`None` 
+        Profile-weighted fraction of the flux from other sources divided
+        by the total flux in g, r and z bands.
+    objtype : :class:`array_like` or :class:`None`
+        The Legacy Surveys TYPE to restrict to point sources.
+    isgaia : :class:`boolean array_like` or :class:`None`
+       ``True`` if there is a match between this object in the Legacy
+       Surveys and in Gaia.
+    pmra, pmdec : :class:`array_like` or :class:`None`
+        Gaia-based proper motion in RA and Dec and parallax
+        (same units as the Gaia data model).
+    aen : :class:`array_like` or :class:`None`
+        Gaia Astrometric Excess Noise (as in the Gaia Data Model).
+    dupsource : :class:`array_like` or :class:`None`
+        Whether the source is a duplicate in Gaia (as in the Gaia Data model).
+    paramssolved : :class:`array_like` or :class:`None`
+        How many parameters were solved for in Gaia (as in the Gaia Data model).
+    primary : :class:`array_like` or :class:`None`
+        ``True`` for objects that should be passed through the selection.
+
+    Returns
+    -------
+    :class:`array_like` 
+        True if and only if the object passes the logic cuts for cmx stars.
+
+    Notes
+    -----
+    - Current version (08/30/18) is version 4 on `the wiki`_.
+    - See also `the Gaia data model`_.
+    """
+    if primary is None:
+        primary = np.ones_like(isgaia, dtype='?')
+        
+    std = primary.copy()
+
+    # ADM A point source with a Gaia match.
+    std &= _psflike(objtype)
+    std &= isgaia
+
+    # ADM An Isolated source.
+    fracflux = [gfracflux, rfracflux, zfracflux]
+    with warnings.catch_warnings():
+        warnings.simplefilter('ignore') # fracflux can be Inf/NaN.
+        for bandint in (0, 1, 2):  # g, r, z.
+            std &= fracflux[bandint] < 0.01
+
+    # ADM No obvious issues with the astrometry.
+    std &= (aen < 1) & (paramssolved == 31)
+
+    # ADM Finite proper motions.
+    std &= np.isfinite(pmra) & np.isfinite(pmdec)
+
+    # ADM Unique source (not a duplicated source).
+    std &= ~dupsource
+
+    return std
+
+
+def isSTD_dither(obs_gflux=None, obs_rflux=None, obs_zflux=None,
+                 isgood=None, primary=None):
+    """Gaia stars for dithering tests during commissioning.
+
+    Parameters
+    ----------
+    obs_gflux, obs_rflux, obs_zflux : :class:`array_like` or :class:`None`
+        The flux in nano-maggies of g, r, z bands WITHOUT any
+        Galactic extinction correction.
+    isgood : :class:`array_like` or :class:`None`
+        ``True`` for objects that pass the logic cuts in
+        :func:`~desitarget.cmx.cmx_cuts.passesSTD_logic`.
+    primary : :class:`array_like` or :class:`None`
+        ``True`` for objects that should be passed through the selection.
+
+    Returns
+    -------
+    :class:`array_like` 
+        True if and only if the object is a Gaia "dither" target.
+
+    Notes
+    -----
+    - Current version (08/30/18) is version 4 on `the wiki`_.
+    - See also `the Gaia data model`_.       
+    """
+    if primary is None:
+        primary = np.ones_like(rflux, dtype='?')
+        
+    isdither = primary.copy()
+    # ADM passes all of the default logic cuts.
+    isdither &= isgood
+
+    # ADM not too bright in g, r, z (> 15 mags)
+    isdither &= obs_gflux < 10**((22.5-15.0)/2.5)
+    isdither &= obs_rflux < 10**((22.5-15.0)/2.5)
+    isdither &= obs_zflux < 10**((22.5-15.0)/2.5)
 
     return isdither
 
 
-def apply_cuts(objects, gaiamatch=False,
-               gaiadir='/project/projectdirs/cosmo/work/gaia/chunks-gaia-dr2-astrom'):
+def isSTD_test(obs_gflux=None, obs_rflux=None, obs_zflux=None,
+               isgood=None, primary=None):
+    """Very bright Gaia stars for early commissioning tests.
+
+    Parameters
+    ----------
+    obs_gflux, obs_rflux, obs_zflux : :class:`array_like` or :class:`None`
+        The flux in nano-maggies of g, r, z bands WITHOUT any
+        Galactic extinction correction.
+    isgood : :class:`array_like` or :class:`None`
+        ``True`` for objects that pass the logic cuts in
+        :func:`~desitarget.cmx.cmx_cuts.passesSTD_logic`.
+    primary : :class:`array_like` or :class:`None`
+        ``True`` for objects that should be passed through the selection.
+
+    Returns
+    -------
+    :class:`array_like` 
+        True if and only if the object is a Gaia "test" target.
+
+    Notes
+    -----
+    - Current version (08/30/18) is version 4 on `the wiki`_.
+    - See also `the Gaia data model`_.       
+    """
+    if primary is None:
+        primary = np.ones_like(rflux, dtype='?')
+        
+    istest = primary.copy()
+    # ADM passes all of the default logic cuts.
+    istest &= isgood
+
+    # ADM not too bright in g, r, z (> 13 mags)
+    istest &= obs_gflux < 10**((22.5-13.0)/2.5)
+    istest &= obs_rflux < 10**((22.5-13.0)/2.5)
+    istest &= obs_zflux < 10**((22.5-13.0)/2.5)
+    # ADM but brighter than dither targets in g (g < 15)
+    istest &= obs_gflux > 10**((22.5-15.0)/2.5)
+
+    return istest
+
+
+def apply_cuts(objects):
     """Perform commissioning (cmx) target selection on objects, return target mask arrays
 
-    Args:
-        objects: numpy structured array with UPPERCASE columns needed for
-            target selection, OR a string tractor/sweep filename
+    Parameters
+    ----------
+    objects: numpy structured array with UPPERCASE columns needed for
+        target selection, OR a string tractor/sweep filename
 
-    Options:
-        gaiamatch : defaults to ``False``
-            if ``True``, match to Gaia DR2 chunks files and populate 
-            Gaia columns to facilitate the MWS selection
-        gaiadir : defaults to the the Gaia DR2 path at NERSC
-             Root directory of a Gaia Data Release as used by the Legacy Surveys. 
+    Returns
+    -------
+    :class:`~numpy.ndarray`
+        commissioning target selection bitmask flags for each object
 
-    Returns:
-        desi_target, where each element is
-        an ndarray of target selection bitmask flags for each object
-
-    See desitarget.cmx.cmx_targetmask for the definition of each bit
+    See desitarget.cmx.cmx_targetmask.cmx_mask for the definition of each bit
     """
     #- Check if objects is a filename instead of the actual data
     if isinstance(objects, str):
         objects = io.read_tractor(objects)
-
-    #ADM add Gaia information, if requested, and if we're going to actually
-    #ADM process the target classes that need Gaia columns
-    if gaiamatch:
-        log.info('Matching Gaia to {} primary objects...t = {:.1f}s'
-                 .format(len(objects),time()-start))
-        gaiainfo = match_gaia_to_primary(objects, gaiadir=gaiadir)
-        log.info('Done with Gaia match for {} primary objects...t = {:.1f}s'
-                 .format(len(objects),time()-start))
-        #ADM remove the GAIA_RA, GAIA_DEC columns as they aren't
-        #ADM in the imaging surveys data model
-        gaiainfo = pop_gaia_coords(gaiainfo)
-        #ADM add the Gaia column information to the primary array
-        for col in gaiainfo.dtype.names:
-            objects[col] = gaiainfo[col]
-
-    #- ensure uppercase column names if astropy Table
+    #- Ensure uppercase column names if astropy Table
     if isinstance(objects, (Table, Row)):
         for col in list(objects.columns.values()):
             if not col.name.isupper():
                 col.name = col.name.upper()
 
-    #ADM flag whether we're using northen (BASS/MZLS) or
-    #ADM southern (DECaLS) photometry
-    photsys_north = _isonnorthphotsys(objects["PHOTSYS"])
-    photsys_south = ~_isonnorthphotsys(objects["PHOTSYS"])
+    # ADM As we need the column names,
+    # ADM capture the case that a single FITS_REC is passed
+    import astropy.io.fits.fitsrec
+    if isinstance(objects, astropy.io.fits.fitsrec.FITS_record):
+        colnames = objects.__dict__['array'].dtype.names
+    else:
+        colnames = objects.dtype.names
 
-    #ADM the observed r-band flux
-    #ADM make copies of values that we may reassign due to NaNs
+    # ADM Currently only coded for objects with Gaia matches
+    # ADM (e.g. DR6 or above). Fail for earlier Data Releases.
+    release = objects['RELEASE']
+    if np.any(release < 6000):
+        log.critical('Commissioning cuts only coded for DR6 or above')
+        raise ValueError
+    if (np.max(objects['PMRA']) == 0.) & np.any(release < 7000):
+        d = "/project/projectdirs/desi/target/gaia_dr2_match_dr6"
+        log.info("Zero objects have a proper motion.")
+        log.critical(
+            "Did you mean to send the Gaia-matched sweeps in, e.g., {}?"
+            .format(d)
+        )
+        raise IOError
+
+    # ADM The observed g/r/z fluxes.
+    obs_gflux = objects['FLUX_G']
     obs_rflux = objects['FLUX_R']
+    obs_zflux = objects['FLUX_Z']
 
-    #- undo Milky Way extinction
+    # ADM The de-extincted g/r/z/ fluxes.
+    from desitarget.cuts import unextinct_fluxes
     flux = unextinct_fluxes(objects)
-
     gflux = flux['GFLUX']
     rflux = flux['RFLUX']
     zflux = flux['ZFLUX']
-    w1flux = flux['W1FLUX']
-    w2flux = flux['W2FLUX']
+
+    # ADM The Legacy Surveys object type and fracflux flags.
     objtype = objects['TYPE']
-    release = objects['RELEASE']
-
-    gfluxivar = objects['FLUX_IVAR_G']
-    rfluxivar = objects['FLUX_IVAR_R']
-    zfluxivar = objects['FLUX_IVAR_Z']
-
-    gnobs = objects['NOBS_G']
-    rnobs = objects['NOBS_R']
-    znobs = objects['NOBS_Z']
-
     gfracflux = objects['FRACFLUX_G']
     rfracflux = objects['FRACFLUX_R']
     zfracflux = objects['FRACFLUX_Z']
 
-    gfracmasked = objects['FRACMASKED_G']
-    rfracmasked = objects['FRACMASKED_R']
-    zfracmasked = objects['FRACMASKED_Z']
-
-    gallmask = objects['ALLMASK_G']
-    rallmask = objects['ALLMASK_R']
-    zallmask = objects['ALLMASK_Z']
-
-    gsnr = objects['FLUX_G'] * np.sqrt(objects['FLUX_IVAR_G'])
-    rsnr = objects['FLUX_R'] * np.sqrt(objects['FLUX_IVAR_R'])
-    zsnr = objects['FLUX_Z'] * np.sqrt(objects['FLUX_IVAR_Z'])
-    w1snr = objects['FLUX_W1'] * np.sqrt(objects['FLUX_IVAR_W1'])
-    w2snr = objects['FLUX_W2'] * np.sqrt(objects['FLUX_IVAR_W2'])
-
-    #ADM issue a warning if gaiamatch was not sent but there's no Gaia information
-    if np.max(objects['PARALLAX']) == 0. and ~gaiamatch:
-        log.warning("Zero objects have a parallax. Did you mean to send gaiamatch?")
-
-    #ADM add the Gaia columns
-    gaia = objects['REF_ID'] != -1
+    # ADM Add the Gaia columns...
+    # ADM if we don't have REF_CAT in the sweeps use the
+    # ADM minimum value of REF_ID to identify Gaia sources. This will
+    # ADM introduce a small number (< 0.001%) of Tycho-only sources.
+    isgaia = objects['REF_ID'] > 0
+    if "REF_CAT" in colnames:
+        isgaia = (objects['REF_CAT'] == b'G2') | (objects['REF_CAT'] == 'G2')
     pmra = objects['PMRA']
     pmdec = objects['PMDEC']
-    parallax = objects['PARALLAX']
-    parallaxivar = objects['PARALLAX_IVAR']
-    #ADM derive the parallax/parallax_error, but set to 0 where the error is bad
-    parallaxovererror = np.where(parallaxivar > 0., parallax*np.sqrt(parallaxivar), 0.)
-    gaiagmag = objects['GAIA_PHOT_G_MEAN_MAG']
-    gaiabmag = objects['GAIA_PHOT_BP_MEAN_MAG']
-    gaiarmag = objects['GAIA_PHOT_RP_MEAN_MAG']
+    pmraivar = objects['PMRA_IVAR']
     gaiaaen = objects['GAIA_ASTROMETRIC_EXCESS_NOISE']
     gaiadupsource = objects['GAIA_DUPLICATED_SOURCE']
 
-    #ADM if the RA proper motion is not NaN, then 31 parameters were solved for
-    #ADM in Gaia astrometry. Use this to set gaiaparamssolved (value is 3 for NaNs)
-    gaiaparamssolved = np.zeros_like(gaia)+31
-    w = np.where(np.isnan(pmra))[0]
-    if len(w) > 0:
-        gaiaparamssolved[w] = 3
+    # ADM If proper motion is not NaN, 31 parameters were solved for
+    # ADM in Gaia astrometry. Or, gaiaparamssolved should be 3 for NaNs).
+    # ADM In the sweeps, NaN has not been preserved...but PMRA_IVAR == 0
+    # ADM in the sweeps is equivalent to PMRA of NaN in Gaia.
+    if 'GAIA_ASTROMETRIC_PARAMS_SOLVED' in colnames:
+        gaiaparamssolved = objects['GAIA_ASTROMETRIC_PARAMS_SOLVED']
+    else:
+        gaiaparamssolved = np.zeros_like(isgaia) + 31
+        w = np.where( np.isnan(pmra) | (pmraivar == 0) )[0]
+        if len(w) > 0:
+            #ADM we need to check the case of a single row being passed
+            if _is_row(gaiaparamssolved):
+                gaiaparamsolved = 3
+            else:
+                gaiaparamssolved[w] = 3
 
-    #ADM test if these columns exist, as they aren't in the Tractor files as of DR7
-    gaiabprpfactor = None
-    gaiasigma5dmax = None
-    try:
-        gaiabprpfactor = objects['GAIA_PHOT_BP_RP_EXCESS_FACTOR']
-        gaiasig5dmax = objects['GAIA_ASTROMETRIC_SIGMA5D_MAX']
-    except:
-        pass
+    # ADM initially, every object passes the cuts (is True).
+    # ADM need to check the case of a single row being passed.
+    if _is_row(objects):
+        primary = np.bool_(True)
+    else:
+        primary = np.ones_like(objects, dtype=bool)
 
-    #ADM Mily Way Selection requires Galactic b
-    _, galb = _gal_coords(objects["RA"],objects["DEC"])
+    # ADM determine if an object passes the default logic for cmx stars
+    isgood = passesSTD_logic(
+        gfracflux=gfracflux, rfracflux=rfracflux, zfracflux=zfracflux,
+        objtype=objtype, isgaia=isgaia, pmra=pmra, pmdec=pmdec,
+        aen=gaiaaen, dupsource=gaiadupsource, paramssolved=gaiaparamssolved,
+        primary=primary
+    )
 
-    dither = isSTD_dither(gflux=gflux)
+    # ADM determine if an object is a "dither" star
+    stddither = isSTD_dither(
+        obs_gflux=obs_gflux, obs_rflux=obs_rflux, obs_zflux=obs_zflux,
+        isgood=isgood, primary=primary
+    )
 
-    # Construct the targetflag bits for DECaLS (i.e. South)
-    # This should really be refactored into a dedicated function.
-    cmx_target  = dither * cmx_mask.STD_GAIA
+    stdtest = isSTD_test(
+        obs_gflux=obs_gflux, obs_rflux=obs_rflux, obs_zflux=obs_zflux,
+        isgood=isgood, primary=primary
+    )
+
+    # ADM Construct the targetflag bits
+    cmx_target  = stddither * cmx_mask.STD_GAIA
+    cmx_target |= stdtest * cmx_mask.STD_TEST
 
     return cmx_target
 
 
-def select_targets(infiles, numproc=4, gaiamatch=False,
-            gaiadir='/project/projectdirs/cosmo/work/gaia/chunks-gaia-dr2-astrom'):
+def select_targets(infiles, numproc=4):
     """Process input files in parallel to select commissioning (cmx) targets
 
     Parameters
     ----------
     infiles : :class:`list` or `str` 
-        A list of input filenames (tractor or sweep files) OR a single filename
+        A list of input filenames (tractor or sweep files) OR a single filename.
     numproc : :class:`int`, optional, defaults to 4
-        The number of parallel processes to use
-    gaiamatch : :class:`boolean`, optional, defaults to ``False``
-        If ``True``, match to Gaia DR2 chunks files and populate Gaia columns
-    gaiadir : :class:`str`, optional, defaults to Gaia DR2 path at NERSC
-        Root directory of a Gaia Data Release as used by the Legacy Surveys.
+        The number of parallel processes to use.
 
     Returns
-    -------   
+    -------
     :class:`~numpy.ndarray`
-        The subset of input targets which pass the cuts, including extra
-        columns for `DESI_TARGET`
+        The subset of input targets which pass the cmx cuts, including an extra
+        column for `CMX_TARGET`.
 
     Notes
     -----
-        - if numproc==1, use serial code instead of parallel
+        - if numproc==1, use serial code instead of parallel.
     """
     from desiutil.log import get_logger
     log = get_logger()
@@ -229,28 +343,28 @@ def select_targets(infiles, numproc=4, gaiamatch=False,
         if not os.path.exists(filename):
             raise ValueError("{} doesn't exist".format(filename))
 
-    def _finalize_targets(objects, desi_target):
+    def _finalize_targets(objects, cmx_target):
         #- desi_target includes BGS_ANY and MWS_ANY, so we can filter just
         #- on desi_target != 0
-        keep = (desi_target != 0)
+        keep = (cmx_target != 0)
         objects = objects[keep]
-        desi_target = desi_target[keep]
+        cmx_target = cmx_target[keep]
 
         #- Add *_target mask columns
-        #ADM note that only desi_target is defined for commissioning
+        #ADM note that only cmx_target is defined for commissioning
         #ADM so just pass that around
-        targets = desitarget.targets.finalize(
-            objects, desi_target, desi_target, desi_target)
+        targets = finalize(objects, cmx_target, cmx_target, cmx_target,
+                           survey='cmx')
 
-        return io.fix_tractor_dr1_dtype(targets)
+        return targets
 
     #- functions to run on every brick/sweep file
     def _select_targets_file(filename):
         '''Returns targets in filename that pass the cuts'''
         objects = io.read_tractor(filename)
-        desi_target, bgs_target, mws_target = apply_cuts(objects,gaiamatch,gaiadir)
+        cmx_target = apply_cuts(objects)
 
-        return _finalize_targets(objects, desi_target, desi_target, desi_target)
+        return _finalize_targets(objects, cmx_target)
 
     # Counter for number of bricks processed;
     # a numpy scalar allows updating nbrick in python 2